refactor: creating user_service to separate handlers from extra logic

+++ backend/internal/models/models.go
@@ -5,7 +5,7 @@ import (
 )
 
 type User struct {
-	ID           int64      `json:"id" db:"id"`
+	ID           uint       `json:"id" db:"id"`
 	FirstName    string     `json:"first_name" db:"first_name"`
 	LastName     string     `json:"last_name" db:"last_name"`
 	Email        string     `json:"email" db:"email"`
